Refactor predict_dataset to a utils folder so it can be used elsewhere
